pyproject.toml
[project]
name = "rpi-aws-iot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "paho-mqtt>=2.1.0",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.0",
    "python-dotenv>=1.0.1",
]
